# src/domain/user.ts
export interface User {
  id: string;
  name: string;
  cpf: string;
  email: string;
  phoneNumber: string;
  createdAt: Date;
  updatedAt: Date;
}

export type CreateUserDto = Omit<User, 'id' | 'createdAt' | 'updatedAt'>; 